intro java

+++ src/main/java/Projek/GoRide/Person.java
@@ -1,19 +1,19 @@
 package Projek.GoRide;
 
 public abstract class Person {
-    private Long id;
+    private int id;
     private String name;
     private int phone;
     private String email;
 
-    public Person(Long id, String name, int phone, String email) {
+    public Person(int id, String name, int phone, String email) {
         this.id = id;
         this.name = name;
         this.phone = phone;
         this.email = email;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
